Unify Data.* and Request.*
I had a few guiding principles here:

1. Organize modules around data structures. How should it be possible to
create these things? If it should only be possible to create these
things through a HTTP request, then that's the only way I'll expose!
Encoding expectations in APIs is awesome.
2. Tailor client data structures to the client, not to the server. For
example, I'm not using updatedAt anywhere. Why decode it, store it, and
expose it just because the server sends it?
3. Unify User, Author, and Profile into User with one type parameter.
Email and AuthToken are only used for the current user, whereas
following is only used for other users. They share all the rest of the
info!
4. Drop Ports.elm in favor of exposing those APIs where they make sense.